Add support for Mod's Grave .WOW files.
Mod's Grave .WOW files are an unusual .MOD variant with a M.K.
signature but 8 channels. They are converted from .669 files using a
converter packaged with the Mod's Grave player for DOS. By calculating
the expected size of a .WOW file, it is possible to detect these files.

This patch also includes some extra checks to avoid false positives in
.WOW detection when loading certain mods (example: ponylips.mod).