Changed XM effect L to behave correctly. (#56)
The XM `Lxx` effect sets the *tick* of the envelopes, but MikMod
previously set the point instead. This predictably lead to bugs of
varying degrees in modules relying on this effect.
See: `Graff/ant attack.xm` on Modland or `villgust.xm` in issue #56
for particularly egregious examples. (Commit message by @AliceLR.)