Add cmake rule to disable shared library builds for libmikmod
The cmakefile has a rule to enable/disable static builds, which is
awesome! However, in some platforms, might not be even possible to
generate a shared object (or at least not trivially). For this, it is
helpful to have a knob to disable said build.

This PR adds a knob for this, which is enabled by default thus not
changing the default behaviour.